# docs/api_specs/ingestion.openapi.yaml
openapi: 3.0.3
info:
  title: RAG Ingestion API
  version: 1.0.0
  description: >-
    FastAPI backend for asynchronous document ingestion into Pinecone using
    Docling + HybridChunker and OpenAI embeddings.
servers:
  - url: https://{host}
    variables:
      host:
        default: api.example.com

paths:
  /v1/ingest:
    post:
      summary: Enqueue documents for ingestion
      description: >-
        Accepts documents for ETL (Docling OCR + HybridChunker), contextualization,
        embeddings, and Pinecone upsert. Processing is asynchronous.
      operationId: ingestDocuments
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IngestMultipartRequest'
            examples:
              single_file:
                summary: Ingest a single file upload
                value:
                  file: handbook.pdf
                  namespace: about_rag
                  index: rag-index
                  metadata_json: '{"document_title":"Company Handbook","source_system":"upload"}'
      responses:
        '202':
          description: Accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestAccepted'
              examples:
                accepted:
                  value:
                    job_id: 9d6f7f8b-6a3d-4b3a-9d0c-1e2a4c8d9f01
                    status: queued
                    received_count: 2
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ingest/{job_id}:
    get:
      summary: Get ingestion job status
      operationId: getIngestStatus
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJobRecord'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    IngestMultipartRequest:
      type: object
      required:
        - file
        - namespace
        - index
      properties:
        file:
          type: string
          format: binary
          description: Uploaded document file (PDF, CSV, JSON, DOCX, or TXT).
        namespace:
          type: string
          description: Pinecone namespace to upsert vectors into (projects, experience, personal, education, about_rag).
        index:
          type: string
          description: Pinecone index name.
        metadata_json:
          type: string
          description: Optional JSON string applied as metadata to the document.
        routing_hint:
          type: string
          description: Optional semantic routing hint for namespace selection.
    IngestAccepted:
      type: object
      required:
        - job_id
        - status
        - received_count
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
        received_count:
          type: integer
          minimum: 1
    IngestJobRecord:
      type: object
      required:
        - job_id
        - filename
        - status
      properties:
        job_id:
          type: string
          format: uuid
        filename:
          type: string
        content_type:
          type: string
          nullable: true
        status:
          type: string
          enum: [queued, processing, completed, failed]
        metadata:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true